Commite by  Raj : Code Changes

+++ src/SchoolManagement.Web/Controllers/ClassSchedulesController.cs
@@ -61,7 +61,10 @@ namespace SchoolManagement.Web.Controllers
                 return View(model);
             }
 
-            int performedBy = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+            if (!TryGetCurrentUserId(out var performedBy))
+            {
+                return Challenge();
+            }
             string ipAddress = HttpContext.Connection.RemoteIpAddress?.ToString() ?? "Unknown";
 
             var result = await _service.SaveAsync(model, performedBy, ipAddress);
@@ -113,7 +116,10 @@ namespace SchoolManagement.Web.Controllers
                 return View(model);
             }
 
-            int performedBy = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+            if (!TryGetCurrentUserId(out var performedBy))
+            {
+                return Challenge();
+            }
             string ipAddress = HttpContext.Connection.RemoteIpAddress?.ToString() ?? "Unknown";
 
             var result = await _service.SaveAsync(model, performedBy, ipAddress);
@@ -143,7 +149,10 @@ namespace SchoolManagement.Web.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            int performedBy = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+            if (!TryGetCurrentUserId(out var performedBy))
+            {
+                return Challenge();
+            }
             string ipAddress = HttpContext.Connection.RemoteIpAddress?.ToString() ?? "Unknown";
 
             var result = await _service.DeleteAsync(id, performedBy, ipAddress);
@@ -166,6 +175,13 @@ namespace SchoolManagement.Web.Controllers
             ViewBag.Classes = new SelectList(classes, "ClassId", "ClassName", model?.ClassId);
             ViewBag.Divisions = new SelectList(divisions, "DivisionId", "DivisionName", model?.DivisionId);
             ViewBag.FinancialYears = new SelectList(financialYears, "FinancialYearId", "FinancialYearName", model?.FinancialYearId);
+        }
+
+        private bool TryGetCurrentUserId(out int userId)
+        {
+            userId = 0;
+            var value = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            return int.TryParse(value, out userId);
         }
     }
 }
